azure-functions/python-functions/requirements.txt
azure-functions-core-tools
azure-functions
Newtonsoft.Json